+++ packages/react/src/_internal/test/mockFile.ts
