+++ packages/react/src/_internal/test/mockIntersectionObserver.ts
