+++ packages/react/src/components/ClientGate/ClientGate.spec.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { screen } from '@testing-library/react';
 
-import { renderSetup } from '../../utils/test/renderSetup';
+import { renderSetup } from '../../_internal/test/renderSetup';
 import { renderToString } from 'react-dom/server';
 
 import { ClientGate } from '.';
