+++ packages/react/src/components/SwitchCase/SwitchCase.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { screen } from '@testing-library/react';
-import { renderSetup } from '../../utils/test/renderSetup';
+import { renderSetup } from '../../_internal/test/renderSetup';
 import { SwitchCase } from '.';
 
 describe('SwitchCase', () => {
