+++ packages/react/src/hooks/useDebouncedInputValue/useDebouncedInputValue.spec.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { screen } from '@testing-library/react';
 import { useDebouncedInputValue } from '.';
-import { renderSetup } from '../../utils/test/renderSetup';
+import { renderSetup } from '../../_internal/test/renderSetup';
 import { delay } from '@modern-kit/utils';
 
 const DELAY = 200;
