+++ packages/react/src/hooks/useMergeRefs/useMergeRefs.spec.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'vitest';
 import { useRef, useState } from 'react';
 import { useMergeRefs } from '.';
 import { screen, waitFor } from '@testing-library/react';
-import { renderSetup } from '../../utils/test/renderSetup';
+import { renderSetup } from '../../_internal/test/renderSetup';
 import { Nullable } from '@modern-kit/types';
 
 const TestComponent = () => {
