+++ packages/react/src/index.ts
@@ -1,3 +1,3 @@
 export * from './components';
 export * from './hooks';
-export * from './types';
+export * from './utils';
